requirements (3).txt
# Core dependencies for CP Chatbot
# Updated to ensure compatibility and remove unused dependencies

# Flask web framework
flask==3.0.0
Werkzeug==3.0.1

# Anthropic Claude AI integration
anthropic==0.8.1

# Environment and configuration management
python-dotenv==1.0.0

# HTTP requests library
requests==2.31.0

# FastCGI support for Dreamhost deployment
flup6==1.1.1

# Flask extensions
flask-cors==4.0.0

# Standard utility libraries
typing-extensions==4.8.0
jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.7.0
click==8.1.7

# Date/time utilities
pytz==2023.3

# Network and certificate handling
idna==3.6
certifi==2023.11.17
charset-normalizer==3.3.2
urllib3==2.1.0

# Optional: For enhanced logging in production
python-json-logger==2.0.7

# Development and testing (optional)
# pytest==7.4.3
# pytest-flask==1.3.0